# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions/{id}": {
      "get": {
        "operationId": "SessionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get session by ID",
        "tags": [
          "sessions"
        ]
      },
      "delete": {
        "operationId": "SessionController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session deleted"
          },
          "404": {
            "description": "Session not found"
          },
          "409": {
            "description": "A credential teardown for the same session name is still in flight (retryable — the body carries `code: 'SESSION_NAME_TEARDOWN_PENDING'`; wait for it to settle and retry), OR another node currently holds this session's live engine and deleting it here would strip a session the owner is running. No destructive side effect runs before either refusal."
          }
        },
        "summary": "Delete a session",
        "tags": [
          "sessions"
        ]
      }
    },
...
    "/api/sessions/{sessionId}/messages": {
      "get": {
        "operationId": "MessageController_getMessages",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "required": false,
            "in": "query",
            "description": "Filter by chat ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Filter by sender. A phone also matches group messages via the author field and any lid that resolves to it.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max messages to return (default 50)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset for pagination",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message history"
          }
        },
        "summary": "Get message history for a session",
        "tags": [
          "messages"
        ]
      }
    },
...
    "/api/stats/sessions/{sessionId}": {
      "get": {
        "operationId": "StatsController_getSessionStats",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Per-session statistics for the requested session."
          }
        },
        "summary": "Get statistics for a specific session",
        "tags": [
          "statistics"
        ]
      }
    },
...
    "/api/sessions/{sessionId}/status": {
      "get": {
        "operationId": "StatusController_getStatuses",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status updates visible to the session, grouped by contact."
          }
        },
        "summary": "Get all contact status updates",
        "tags": [
          "status"
        ]
      }
    },
    "/api/sessions/{sessionId}/status/{contactId}": {
      "get": {
        "operationId": "StatusController_getContactStatus",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status updates from the requested contact."
          }
        },
        "summary": "Get status updates from a specific contact",
        "tags": [
          "status"
        ]
      }
    },
    "/api/sessions/{sessionId}/status/{statusId}/media": {
      "get": {
        "operationId": "StatusController_getStatusMedia",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statusId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The status image/video bytes."
          },
          "404": {
            "description": "No stored media (text status, omitted, or expired)."
          }
        },
        "summary": "Stream a stored status media file",
        "tags": [
          "status"
        ]
      }
    },
    "/api/sessions/{sessionId}/status/send-text": {
      "post": {
        "operationId": "StatusController_sendTextStatus",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendTextStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Text status posted. The recipients allow-list is honored on Baileys only; whatsapp-web.js broadcasts to the account's status-privacy audience."
          },
          "400": {
            "description": "Invalid request, or the post was blocked by a plugin."
          }
        },
        "summary": "Post a text status",
        "tags": [
          "status"
        ]
      }
    },
    "/api/sessions/{sessionId}/status/send-image": {
      "post": {
        "operationId": "StatusController_sendImageStatus",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendImageStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image status posted. The recipients allow-list is honored on Baileys only; whatsapp-web.js broadcasts to the account's status-privacy audience."
          },
          "400": {
            "description": "Neither url nor base64 provided, or the post was blocked by a plugin."
          },
          "413": {
            "description": "Base64 media exceeds MEDIA_DOWNLOAD_MAX_BYTES."
          }
        },
        "summary": "Post an image status",
        "tags": [
          "status"
        ]
      }
    },
    "/api/sessions/{sessionId}/status/send-video": {
      "post": {
        "operationId": "StatusController_sendVideoStatus",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendVideoStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Video status posted. The recipients allow-list is honored on Baileys only; whatsapp-web.js broadcasts to the account's status-privacy audience."
          },
          "400": {
            "description": "Neither url nor base64 provided, or the post was blocked by a plugin."
          },
          "413": {
            "description": "Base64 media exceeds MEDIA_DOWNLOAD_MAX_BYTES."
          }
        },
        "summary": "Post a video status",
        "tags": [
          "status"
        ]
      }
    },
    "/api/sessions/{sessionId}/status/send-voice": {
      "post": {
        "operationId": "StatusController_sendVoiceStatus",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendVoiceStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Voice status posted. WhatsApp plays a status voice note only as Ogg/Opus and neither engine transcodes, so convert first via POST /media/convert/voice. The recipients allow-list is honored on Baileys only; whatsapp-web.js broadcasts to the account's status-privacy audience."
          },
          "400": {
            "description": "Neither url nor base64 provided, or the post was blocked by a plugin."
          },
          "413": {
            "description": "Base64 media exceeds MEDIA_DOWNLOAD_MAX_BYTES."
          }
        },
        "summary": "Post an audio status as a voice note",
        "tags": [
          "status"
        ]
      }
    },
    "/api/sessions/{sessionId}/status/{statusId}": {
      "delete": {
        "operationId": "StatusController_deleteStatus",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statusId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status deleted."
          }
        },
        "summary": "Delete own status",
        "tags": [
          "status"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.14.2",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [
    {
      "url": "/",
      "description": "This instance (the origin serving these docs)"
    },
    {
      "url": "http://{host}:{port}",
      "description": "Another instance (set host and port)",
      "variables": {
        "host": {
          "default": "localhost"
        },
        "port": {
          "default": "2785",
          "description": "PORT env var"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "metrics-bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "METRICS_TOKEN for the GET /api/metrics scrape endpoint"
      }
    },
    "schemas": {
...
      "AccountRestrictionDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "reachout_timelock",
              "tos_block",
              "proxy_block"
            ],
            "description": "What WhatsApp is restricting. `reachout_timelock` leaves the session connected and existing chats working, blocking only the start of new conversations. `tos_block` and `proxy_block` are connection-level refusals — the session cannot stay linked while one is in force, so seeing either alongside a `ready` status is not possible.",
            "example": "reachout_timelock"
          },
          "code": {
            "type": "string",
            "description": "The engine's own token for the cause, passed through verbatim so it can be searched for and so a value newer than this gateway is still surfaced rather than flattened.",
            "example": "BIZ_QUALITY"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "When enforcement ends, if the engine states it. Only reachout timelocks carry an expiry; absent means the engine gave no end time, not that the restriction is permanent.",
            "example": "2026-08-04T09:00:00Z",
            "nullable": true
          }
        },
        "required": [
          "kind",
          "code"
        ]
      },
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "sess_123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "example": "my-bot"
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "initializing",
              "qr_ready",
              "authenticating",
              "ready",
              "disconnected",
              "action_required",
              "failed"
            ],
            "example": "ready"
          },
          "phone": {
            "type": "string",
            "example": "628123456789",
            "nullable": true
          },
          "pushName": {
            "type": "string",
            "example": "John Doe",
            "nullable": true
          },
          "connectedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-02-02T10:00:00Z",
            "nullable": true
          },
          "lastActive": {
            "type": "string",
            "format": "date-time",
            "example": "2025-02-02T10:30:00Z",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-02-02T09:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-02-02T10:00:00Z"
          },
          "lastError": {
            "type": "string",
            "description": "Human-readable reason carried while the status is FAILED (a terminal engine failure) or ACTION_REQUIRED (the engine is running but something needs a human). Cleared on any other status.",
            "example": "Failed to launch the browser process: spawn /usr/bin/chromium ENOENT",
            "nullable": true
          },
          "restriction": {
            "description": "A restriction WhatsApp itself has placed on this session's account, or null when there is none. Distinct from `lastError`, which describes a fault on our side of the link. Derived from live engine state, so it is never persisted: it is re-established on the next connect.",
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountRestrictionDto"
              }
            ]
          },
          "engineLoaded": {
            "type": "boolean",
            "description": "Whether the gateway currently holds a live engine for this session. This is the precondition the lifecycle routes actually enforce, and `status` alone does not imply it: a `disconnected` session keeps its engine for the duration of an automatic reconnect backoff, while a session stopped through `POST /sessions/:id/stop` carries the same status with no engine. When `true`, `stop`, `logout` and `force-kill` can act and `start` answers 400; when `false`, the reverse. Derived per request from live process state, so it is never persisted and never historical.",
            "example": true
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "createdAt",
          "updatedAt",
          "engineLoaded"
        ]
      },
...
      "SendTextStatusDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Status text body.",
            "example": "Out for delivery 📦",
            "maxLength": 4096
          },
          "backgroundColor": {
            "type": "string",
            "description": "Background color (hex).",
            "example": "#25D366"
          },
          "font": {
            "type": "number",
            "description": "Font family index from the WhatsApp status font enum: 0 (default), 1, 2, 6 (bold), 7, 8, 9, or 10. whatsapp-web.js honors only 0–7 and clamps anything above to the default.",
            "example": 0,
            "enum": [
              0,
              1,
              2,
              6,
              7,
              8,
              9,
              10
            ]
          },
          "recipients": {
            "description": "Recipient JIDs (0–256). WhatsApp Status is not posted to a group — use @c.us or @lid individuals. Required on the Baileys engine (it posts to exactly this allow-list); ignored by whatsapp-web.js, which broadcasts to the account's status-privacy audience — omit it there.",
            "example": [
              "628123456789@c.us"
            ],
            "maxItems": 256,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "text"
        ]
      },
      "StatusMediaInput": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "Public http(s) URL of the media (server-fetched, SSRF-guarded).",
            "example": "https://example.com/banner.jpg"
          },
          "base64": {
            "type": "string",
            "description": "Base64-encoded media. Requires mimetype.",
            "example": "/9j/4AAQSkZJRg..."
          },
          "mimetype": {
            "type": "string",
            "description": "MIME type. Required when sending base64.",
            "example": "image/jpeg"
          }
        }
      },
      "SendImageStatusDto": {
        "type": "object",
        "properties": {
          "image": {
            "description": "Image source (URL or base64).",
            "allOf": [
              {
                "$ref": "#/components/schemas/StatusMediaInput"
              }
            ]
          },
          "caption": {
            "type": "string",
            "description": "Optional caption.",
            "example": "New drop!",
            "maxLength": 1024
          },
          "recipients": {
            "description": "Recipient JIDs (0–256), @c.us or @lid. Required on the Baileys engine (it posts to exactly this allow-list); ignored by whatsapp-web.js, which broadcasts to the account's status-privacy audience — omit it there.",
            "example": [
              "628123456789@c.us"
            ],
            "maxItems": 256,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "image"
        ]
      },
      "SendVideoStatusDto": {
        "type": "object",
        "properties": {
          "video": {
            "description": "Video source (URL or base64).",
            "allOf": [
              {
                "$ref": "#/components/schemas/StatusMediaInput"
              }
            ]
          },
          "caption": {
            "type": "string",
            "description": "Optional caption.",
            "example": "Demo",
            "maxLength": 1024
          },
          "recipients": {
            "description": "Recipient JIDs (0–256), @c.us or @lid. Required on the Baileys engine (it posts to exactly this allow-list); ignored by whatsapp-web.js, which broadcasts to the account's status-privacy audience — omit it there.",
            "example": [
              "628123456789@c.us"
            ],
            "maxItems": 256,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "video"
        ]
      },
      "SendVoiceStatusDto": {
        "type": "object",
        "properties": {
          "audio": {
            "description": "Audio source (URL or base64). WhatsApp plays a status voice note only as Ogg/Opus, and neither engine transcodes — use the media conversion endpoint to produce it. The mimetype defaults to 'audio/ogg; codecs=opus'.",
            "allOf": [
              {
                "$ref": "#/components/schemas/StatusMediaInput"
              }
            ]
          },
          "backgroundColor": {
            "type": "string",
            "description": "Background colour as `#RRGGBB`, which WhatsApp renders behind the voice-note bubble. Baileys only — whatsapp-web.js exposes no styling for a status and ignores it.",
            "example": "#25D366"
          },
          "recipients": {
            "description": "Recipient JIDs (0–256), @c.us or @lid. Required on the Baileys engine (it posts to exactly this allow-list); ignored by whatsapp-web.js, which broadcasts to the account's status-privacy audience — omit it there.",
            "example": [
              "628123456789@c.us"
            ],
            "maxItems": 256,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "audio"
        ]
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}